Refactor module to only include available options
