+++ README.md
@@ -1,13 +1,20 @@
+# n-brightcove-to-uuid
 
-Attempt to generate a UUID v3 (MD5) from a Brightcove ID,
+Generates a MD5 UUID from a given Brightcove 11 or 13 digit ID.
 
+## Installation
+
+```sh
+# install from NPM
+$ npm i -S @financial-times/n-brightcove-to-uuid
 ```
-npm install
-node test
-```
 
-Feed a list of Brightcove IDs to the generator, 
+## Usage
 
-```
-find ../next-video-archive/archive -type f | cut -d '/' -f 4 | node generate-uuids-from-brightcove
+```js
+const nBrightcoveToUuid = require('@financial-times/n-brightcove-to-uuid');
+
+const id = nBrightcoveToUuid('834961745001');
+
+console.log(id); // 564b8be3-0775-3531-b87b-f6da2272f642
 ```
