+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@financial-times/n-brightcove-to-uuid",
   "version": "0.0.0",
-  "main": "src/v3",
+  "main": "index.js",
   "scripts": {
     "test": "tap test"
   },
